[#44]Refactor: DimView base view에 추가
- DimViewManager 삭제
- self를 참조해서 hide를 하다보니 뷰 안에서만 hide를 할 수 있어서
- base에 포함하게 됨.
